+++ src/q_element.js
@@ -11,7 +11,27 @@ var Me = module.exports = function QElement(domElement, description) {
 	this._domElement = domElement;
 	this._description = description;
 
-//	this.top = ElementEdge.top(this);
+	this.top = ElementEdge.top(this);
+	this.right = ElementEdge.right(this);
+	this.bottom = ElementEdge.bottom(this);
+	this.left = ElementEdge.left(this);
+};
+
+Me.prototype.diff = function(expected) {
+	ensure.signature(arguments, [ Object ]);
+
+	var result = [];
+	var keys = objectKeys(expected);
+	var key, diff, constraint;
+	for (var i = 0; i < keys.length; i++) {
+		key = keys[i];
+		constraint = this[key];
+		ensure.that(constraint !== undefined, "'" + key + "' is unknown and can't be used with diff()");
+		diff = constraint.diff(expected[key]);
+		if (diff !== "") result.push(diff);
+	}
+
+	return result.join("\n");
 };
 
 Me.prototype.getRawStyle = function(styleName) {
@@ -70,3 +90,43 @@ Me.prototype.equals = function(that) {
 
 	return this._domElement === that._domElement;
 };
+
+// WORKAROUND IE8: No Object.keys
+function objectKeys(obj) {
+	if (Object.keys) return Object.keys(obj);
+
+	// From https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/keys
+  var hasOwnProperty = Object.prototype.hasOwnProperty,
+      hasDontEnumBug = !({ toString: null }).propertyIsEnumerable('toString'),
+      dontEnums = [
+        'toString',
+        'toLocaleString',
+        'valueOf',
+        'hasOwnProperty',
+        'isPrototypeOf',
+        'propertyIsEnumerable',
+        'constructor'
+      ],
+      dontEnumsLength = dontEnums.length;
+
+  if (typeof obj !== 'object' && (typeof obj !== 'function' || obj === null)) {
+    throw new TypeError('Object.keys called on non-object');
+  }
+
+  var result = [], prop, i;
+
+  for (prop in obj) {
+    if (hasOwnProperty.call(obj, prop)) {
+      result.push(prop);
+    }
+  }
+
+  if (hasDontEnumBug) {
+    for (i = 0; i < dontEnumsLength; i++) {
+      if (hasOwnProperty.call(obj, dontEnums[i])) {
+        result.push(dontEnums[i]);
+      }
+    }
+  }
+  return result;
+}